pyproject.toml
[tool.poetry]
name = "note-service"
version = "0.1.0"
description = ""
authors = [" annavaleriev <filenko.a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
